:truck: Moves sort params and enum
