+++ src/types/index.ts
@@ -1,4 +1,4 @@
-import { QueryParams } from './queryParams/queryParams'
+import { QueryParams } from './queryParams.types.'
 import { FetchDataServiceInterface } from './service'
 
 export { QueryParams, FetchDataServiceInterface }
