--- src/types/queryParams/sort.ts
@@ -1,9 +1,0 @@
-enum SortOrderEnum {
-    ASC = `asc`,
-    DESC = `desc`,
-}
-
-export type Sort = {
-    sortOrder: SortOrderEnum
-    sortField: string
-}
